requirements.txt
annotated-types==0.7.0
asyncio==4.0.0
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.6
lxml==6.0.2
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.2.0
plum-dispatch==1.7.4
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pypdfium2==5.6.0
python-docx==1.2.0
spire-doc==14.1.6
typing-inspection==0.4.2
typing_extensions==4.15.0
